requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.3
requests==2.31.0
pydantic==2.8.2
PyYAML==6.0.1
click==8.1.7
rich==13.7.0